+++ indico/modules/legal/controllers.py
@@ -5,14 +5,16 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
-from flask import redirect
+from flask import redirect, session
 
 from indico.modules.admin import RHAdminBase
-from indico.modules.legal import legal_settings
-from indico.modules.legal.forms import LegalMessagesForm
-from indico.modules.legal.views import WPDisplayPrivacyPolicy, WPDisplayTOS, WPManageLegalMessages
+from indico.modules.legal import check_terms_required, legal_settings
+from indico.modules.legal.forms import LegalMessagesForm, create_agreement_form
+from indico.modules.legal.views import WPDisplayAgreement, WPDisplayPrivacyPolicy, WPDisplayTOS, WPManageLegalMessages
+from indico.util.date_time import now_utc
 from indico.web.flask.util import url_for
 from indico.web.rh import RH
+from indico.web.util import url_for_index
 
 
 class RHManageLegalMessages(RHAdminBase):
@@ -38,3 +40,17 @@ class RHDisplayPrivacyPolicy(RH):
         if url:
             return redirect(url)
         return WPDisplayPrivacyPolicy.render_template('privacy.html', content=legal_settings.get('privacy_policy'))
+
+
+class RHAcceptTerms(RH):
+    def _process(self):
+        if not check_terms_required():
+            return redirect(url_for_index())
+
+        form = create_agreement_form()
+        if form.validate_on_submit():
+            if form.accept_terms.data:
+                session.user.accepted_terms_dt = now_utc()
+            return redirect(session.pop('legal_agreement_return_path', url_for_index()))
+
+        return WPDisplayAgreement.render_template('agreement.html', form=form, **legal_settings.get_all())