+++ CHANGES.rst
@@ -85,6 +85,8 @@ Improvements
   ``indico.conf`` setting) (:pr:`6028`, thanks :user:`openprojects`)
 - Add option to exceptionally grant registration modification privileges to some registrants
   (:issue:`5264`, :pr:`6152`, thanks :user:`Thanhphan1147`)
+- Add option to require users to agree to terms during signup or after they have been updated
+  (:issue:`5923`, :pr:`5925`, thanks :user:`kewisch`)
 
 Bugfixes
 ^^^^^^^^